2018: Day 9/Part 2
I had an implementation from last year that I tossed away (because it
seemed it wasn't going to improve much over the part1 implementation).
I was going to rewrite it using a doubly linked list to avoid memcpy'ing on
each insert and removal (as a long shot, as I wasn't confident that alone would
be enough), and while I was searching whether the stdlib had such an
implementation (container/list) I stumbled across container/ring, which is
exactly what I needed! I didn't expect the improvements to be this dramatic
(specially with this kind of data structure), but it clearly paid off,
and I learned something new about Golang's container package.